# apps/backend/app/api/routers/jobs.py
"""Processing job transport endpoints."""

from fastapi import (
    APIRouter,
    File,
    Form,
    Header,
    HTTPException,
    Query,
    UploadFile,
    status,
)
from fastapi.responses import Response
from starlette.concurrency import run_in_threadpool

from app.api.administrative_access import require_administrator
from app.api.dependencies import (
    JobMutationConflictError,
    JobTransportNotFoundError,
    JobUploadConflictError,
    JobUploadInputError,
    JobUploadParserConfigurationError,
    JobUploadParserProviderError,
    JobUploadUnexpectedParserError,
)
from app.api.response_contracts import SUPPORTED_IMAGE_RESPONSE_CONTENT
from app.application.jobs import (
    JobMutationService,
    JobQueryService,
    JobUploadPipelineRequest,
    JobUploadRequest,
    JobUploadService,
)
from app.application.admin_ocr_test import AuthorizeAdministrator
from app.domain.poker import CanonicalState
from app.domain.hands import JobQueue, JobRecord, ScreenshotMetadataRequest


def create_jobs_router(
    runtime: JobQueryService,
    authorize_administrator: AuthorizeAdministrator,
) -> APIRouter:
    """Build the processing job read router with application dependencies."""

    router = APIRouter()

    @router.get(
        "/api/admin/ocr/jobs",
        operation_id="admin_ocr_jobs_list",
        response_model=JobQueue,
    )
    def get_processing_jobs(
        limit: int = Query(default=100, ge=1, le=100),
        offset: int = Query(default=0, ge=0),
        authorization: str | None = Header(
            default=None,
            alias="Authorization",
            include_in_schema=False,
        ),
    ) -> JobQueue:
        require_administrator(authorize_administrator(authorization))
        return runtime.list_jobs(limit, offset)

    @router.get(
        "/api/admin/ocr/jobs/{job_id}",
        operation_id="admin_ocr_job_get",
        response_model=JobRecord,
    )
    def get_job(
        job_id: str,
        authorization: str | None = Header(
            default=None,
            alias="Authorization",
            include_in_schema=False,
        ),
    ) -> JobRecord:
        require_administrator(authorize_administrator(authorization))
        try:
            return runtime.get_job(job_id)
        except JobTransportNotFoundError as exc:
            raise HTTPException(status_code=404, detail=str(exc)) from exc

    @router.get(
        "/api/admin/ocr/jobs/{job_id}/image",
        operation_id="admin_ocr_job_image_get",
        response_class=Response,
        responses={"200": {"content": SUPPORTED_IMAGE_RESPONSE_CONTENT}},
    )
    def get_job_image(
        job_id: str,
        authorization: str | None = Header(
            default=None,
            alias="Authorization",
            include_in_schema=False,
        ),
    ) -> Response:
        require_administrator(authorize_administrator(authorization))
        try:
            image = runtime.get_image(job_id)
        except JobTransportNotFoundError as exc:
            raise HTTPException(status_code=404, detail=str(exc)) from exc
        return Response(content=image.content, media_type=image.media_type)

    return router


def create_job_upload_router(runtime: JobUploadService) -> APIRouter:
    """Build the multipart processing-job upload router (administrative OCR test surface)."""

    router = APIRouter()

    @router.post(
        "/api/admin/ocr/jobs",
        operation_id="admin_ocr_jobs_create",
        response_model=JobRecord,
        status_code=status.HTTP_201_CREATED,
    )
    async def create_job(
        file: UploadFile = File(...),
        upload_request_id: str | None = Form(
            default=None,
            min_length=1,
            max_length=128,
            pattern=r"^[A-Za-z0-9._:-]+$",
        ),
        parser_provider: str | None = Form(
            default=None,
            min_length=1,
            max_length=64,
            pattern=r"^[a-z0-9_]+$",
        ),
        parser_layout_profile: str | None = Form(
            default=None,
            min_length=1,
            max_length=64,
            pattern=r"^[a-z0-9_]+$",
        ),
        authorization: str | None = Header(
            default=None,
            alias="Authorization",
            include_in_schema=False,
        ),
    ) -> JobRecord:
        require_administrator(runtime.authorize_administrator(authorization))

        pipeline_request = JobUploadPipelineRequest(
            parser_provider=parser_provider,
            parser_layout_profile=parser_layout_profile,
        )
        try:
            selection = runtime.resolve_pipeline(pipeline_request)
        except JobUploadInputError as exc:
            raise HTTPException(status_code=400, detail=str(exc)) from exc

        image_bytes = await file.read(runtime.max_upload_bytes + 1)
        if len(image_bytes) > runtime.max_upload_bytes:
            raise HTTPException(status_code=413, detail="Upload exceeds maximum size")

        request = JobUploadRequest(
            original_filename=file.filename or "screenshot.png",
            image_bytes=image_bytes,
            upload_request_id=upload_request_id,
            selection=selection,
        )
        try:
            return await run_in_threadpool(runtime.process_upload, request)
        except JobUploadInputError as exc:
            raise HTTPException(status_code=400, detail=str(exc)) from exc
        except JobUploadConflictError as exc:
            raise HTTPException(status_code=409, detail=str(exc)) from exc
        except JobUploadParserConfigurationError as exc:
            raise HTTPException(
                status_code=500,
                detail=f"Parser configuration error: {exc}",
            ) from exc
        except JobUploadParserProviderError as exc:
            raise HTTPException(status_code=502, detail=str(exc)) from exc
        except JobUploadUnexpectedParserError as exc:
            raise HTTPException(status_code=500, detail=str(exc)) from exc

    return router


def create_job_mutations_router(
    runtime: JobMutationService,
    authorize_administrator: AuthorizeAdministrator,
) -> APIRouter:
    """Build the processing job mutation router with application dependencies."""

    router = APIRouter()

    @router.put(
        "/api/admin/ocr/jobs/{job_id}/metadata",
        operation_id="admin_ocr_job_metadata_update",
        response_model=JobRecord,
    )
    def update_job_metadata(
        job_id: str,
        metadata: ScreenshotMetadataRequest,
        authorization: str | None = Header(
            default=None,
            alias="Authorization",
            include_in_schema=False,
        ),
    ) -> JobRecord:
        require_administrator(authorize_administrator(authorization))
        try:
            return runtime.update_metadata(job_id, metadata)
        except JobTransportNotFoundError as exc:
            raise HTTPException(status_code=404, detail=str(exc)) from exc

    @router.delete(
        "/api/admin/ocr/jobs/{job_id}",
        operation_id="admin_ocr_job_delete",
        status_code=status.HTTP_204_NO_CONTENT,
        response_class=Response,
    )
    def delete_job(
        job_id: str,
        authorization: str | None = Header(
            default=None,
            alias="Authorization",
            include_in_schema=False,
        ),
    ) -> Response:
        require_administrator(authorize_administrator(authorization))
        try:
            runtime.delete_job(job_id)
        except JobTransportNotFoundError as exc:
            raise HTTPException(status_code=404, detail=str(exc)) from exc
        except JobMutationConflictError as exc:
            raise HTTPException(status_code=409, detail=str(exc)) from exc
        return Response(status_code=status.HTTP_204_NO_CONTENT)

    @router.post(
        "/api/admin/ocr/jobs/{job_id}/approve",
        operation_id="admin_ocr_job_approve",
        response_model=JobRecord,
    )
    def approve_job(
        job_id: str,
        state: CanonicalState,
        authorization: str | None = Header(
            default=None,
            alias="Authorization",
            include_in_schema=False,
        ),
    ) -> JobRecord:
        require_administrator(authorize_administrator(authorization))
        try:
            return runtime.approve_job(job_id, state)
        except JobTransportNotFoundError as exc:
            raise HTTPException(status_code=404, detail=str(exc)) from exc

    return router