# packages/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ArchiveJobsRequest": {
        "properties": {
          "job_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Job Ids",
            "type": "array"
          }
        },
        "required": ["job_ids"],
        "title": "ArchiveJobsRequest",
        "type": "object"
      },
...
      "Body_admin_ocr_jobs_create": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "parser_layout_profile": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "pattern": "^[a-z0-9_]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Layout Profile"
          },
          "parser_provider": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "pattern": "^[a-z0-9_]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Provider"
          },
          "upload_request_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9._:-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Request Id"
          }
        },
        "required": ["file"],
        "title": "Body_admin_ocr_jobs_create",
        "type": "object"
      },
      "CanonicalState": {
        "properties": {
          "action_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Context"
          },
          "board_cards": {
            "items": {
              "$ref": "#/components/schemas/Card"
            },
            "title": "Board Cards",
            "type": "array"
          },
          "completed_postflop_streets": {
            "items": {
              "$ref": "#/components/schemas/CompletedPostflopStreetHistory"
            },
            "maxItems": 2,
            "title": "Completed Postflop Streets",
            "type": "array"
          },
          "current_bet": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Bet"
          },
          "effective_stack": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Stack"
          },
          "facing_action": {
            "anyOf": [
              {
                "enum": ["bet", "raise"],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facing Action"
          },
          "hero_cards": {
            "items": {
              "$ref": "#/components/schemas/Card"
            },
            "title": "Hero Cards",
            "type": "array"
          },
          "hero_position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hero Position"
          },
          "hero_stack": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hero Stack"
          },
          "opponent_commitment_total": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Commitment Total"
          },
          "opponent_position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Position"
          },
          "opponent_stack": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Stack"
          },
          "opponent_wager": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Wager"
          },
          "opponents_at_current_bet": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponents At Current Bet"
          },
          "players_in_hand": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Players In Hand"
          },
          "postflop_action_history": {
            "items": {
              "$ref": "#/components/schemas/PostflopAction"
            },
            "maxItems": 8,
            "title": "Postflop Action History",
            "type": "array"
          },
          "pot_size": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pot Size"
          },
          "preflop_action_history": {
            "items": {
              "$ref": "#/components/schemas/PreflopAction"
            },
            "maxItems": 8,
            "title": "Preflop Action History",
            "type": "array"
          },
          "preflop_open_size": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preflop Open Size"
          },
          "preflop_opener_position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preflop Opener Position"
          },
          "street": {
            "anyOf": [
              {
                "enum": ["preflop", "flop", "turn", "river"],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Street"
          },
          "user_approved": {
            "default": false,
            "title": "User Approved",
            "type": "boolean"
          }
        },
        "title": "CanonicalState",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobHistory": {
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobRecord"
            },
            "title": "Jobs",
            "type": "array"
          },
          "snapshot_version": {
            "title": "Snapshot Version",
            "type": "string"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": ["total", "snapshot_version"],
        "title": "JobHistory",
        "type": "object"
      },
      "JobQueue": {
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobRecord"
            },
            "title": "Jobs",
            "type": "array"
          },
          "snapshot_version": {
            "title": "Snapshot Version",
            "type": "string"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": ["total", "snapshot_version"],
        "title": "JobQueue",
        "type": "object"
      },
      "JobRecord": {
        "properties": {
          "approved_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CanonicalState"
              },
              {
                "type": "null"
              }
            ]
          },
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "benchmark_import_request_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Benchmark Import Request Id"
          },
          "benchmark_included": {
            "default": false,
            "title": "Benchmark Included",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "image_filename": {
            "title": "Image Filename",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "original_filename": {
            "title": "Original Filename",
            "type": "string"
          },
          "parser_auto_approval_eligible": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Auto Approval Eligible"
          },
          "parser_layout_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Layout Profile"
          },
          "parser_provider": {
            "title": "Parser Provider",
            "type": "string"
          },
          "parser_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParserResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "default": "created",
            "enum": ["created", "parsed", "approved", "error"],
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "upload_request_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9._:-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Request Id"
          }
        },
        "required": ["original_filename", "image_filename", "parser_provider"],
        "title": "JobRecord",
        "type": "object"
      },
...
      "ScreenshotMetadataRequest": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "ScreenshotMetadataRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Poker Training Analyzer API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/ocr/history": {
      "get": {
        "operationId": "admin_ocr_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobHistory"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get History"
      },
      "put": {
        "operationId": "admin_ocr_history_archive",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveJobsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobHistory"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Jobs"
      }
    },
    "/api/admin/ocr/jobs": {
      "get": {
        "operationId": "admin_ocr_jobs_list",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobQueue"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Processing Jobs"
      },
      "post": {
        "operationId": "admin_ocr_jobs_create",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_admin_ocr_jobs_create"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job"
      }
    },
    "/api/admin/ocr/jobs/{job_id}": {
      "delete": {
        "operationId": "admin_ocr_job_delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Job"
      },
      "get": {
        "operationId": "admin_ocr_job_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job"
      }
    },
    "/api/admin/ocr/jobs/{job_id}/approve": {
      "post": {
        "operationId": "admin_ocr_job_approve",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CanonicalState"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Job"
      }
    },
...
    "/api/admin/ocr/jobs/{job_id}/image": {
      "get": {
        "operationId": "admin_ocr_job_image_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/gif": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/jpeg": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/png": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/webp": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Image"
      }
    },
    "/api/admin/ocr/jobs/{job_id}/metadata": {
      "put": {
        "operationId": "admin_ocr_job_metadata_update",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreenshotMetadataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Job Metadata"
      }
    },
...
  }
}